refactor(orchestrator): move consolidation schemas to orchestrator-config, delete run-config-v2
Moves ProviderConfigSchema, SecurityConfigInputSchema, and
BrainConfigSchema from the dead run-config-v2.ts into
orchestrator-config.ts where they are actually consumed.
Deletes run-config-v2.ts and its test file.
